# Repository.openapi+json.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "Repository API",
        "description": "",
        "version": "1.0"
    },
    "paths": {
        "/v1/admin-actions/{adminActionId}": {
            "get": {
                "tags": [
                    "AdminActions"
                ],
                "parameters": [
                    {
                        "name": "adminActionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "AdminActions"
                ],
                "parameters": [
                    {
                        "name": "adminActionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "AdminActions"
                ],
                "parameters": [
                    {
                        "name": "adminActionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/admin-actions": {
            "get": {
                "tags": [
                    "AdminActions"
                ],
                "parameters": [
                    {
                        "name": "gameType",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/GameType"
                        }
                    },
                    {
                        "name": "playerId",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "adminId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/AdminActionFilter"
                        }
                    },
                    {
                        "name": "skipEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "takeEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/AdminActionOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "AdminActions"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminActionFilter": {
                "enum": [
                    "ActiveBans",
                    "UnclaimedBans"
                ],
                "type": "string",
                "format": "int32"
            },
            "AdminActionOrder": {
                "enum": [
                    "CreatedAsc",
                    "CreatedDesc"
                ],
                "type": "string",
                "format": "int32"
            },
...
            "GameType": {
                "enum": [
                    "Unknown",
                    "CallOfDuty2",
                    "CallOfDuty4",
                    "CallOfDuty5",
                    "Insurgency",
                    "ArkSurvivalEvolved",
                    "Battlefield1",
                    "Battlefield3",
                    "Battlefield4",
                    "Battlefield5",
                    "BattlefieldBadCompany2",
                    "CrysisWars",
                    "Left4Dead2",
                    "Minecraft",
                    "PlayerUnknownsBattleground",
                    "RisingStormVietnam",
                    "Rust",
                    "WarThunder",
                    "WorldOfWarships",
                    "WorldWar3",
                    "UnrealTournament2004",
                    "Arma",
                    "Arma2",
                    "Arma3"
                ],
                "type": "string",
                "format": "int32"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}